requirements.txt
feedparser==6.0.2
python-telegram-bot==20.3
requests==2.26.0
pymongo

